[BROKEN] Multiple substitutions, this is wrong
Make is a tool to organize the build of a whole project. Thus, there are
no files outside of that project.
If there are, then the only way to handle this is by setting the VPATH.

See and READ:
http://make.paulandlesley.org/
And the paper "Recursive make considered harmful"

(yes, all of it)